MAT-7941: Demote Terminology errors to warnings during test case validation for QI-Core 6 only by replacing the In Memory Terminology validator with our custom lenient validator, composed of the In Memory validator.
The In Memory validator is unable to expand the internal US-Core value sets (because they now point to VSAC instead of being a static list of codes) and concludes codes used in test cases are invalid.

Accept the result of the In Memory Validator and demote any Error findings to Warning.